SCHED-95: Modify literal string with GeminiProperties enum. Also decorator and GeminiObservation Class are in api.observatory.gemini to 1. Keep all specific observatory code in one place and 2. To remove circular dependencies
