+++ common/decorators/__init__.py
@@ -1,6 +1,5 @@
 
 import numpy as np
-from datetime import timedelta
 
 def scalar_input(func):
     """
@@ -17,15 +16,3 @@ def scalar_input(func):
            
         return np.squeeze(func(*args, **kwargs))
     return wrapper
-
-
-def with_igrins_cal(func):
-    def add_calibration(self):
-        if ('IGRINS' in [res.id for res in self.required_resources()] and self.partner_used() > 0):
-            return func(self) + timedelta(seconds=(1 / 6))
-        return func(self)
-    return add_calibration
-
-"""
-def gemini(func)
-"""
